Sample Classes
These are a couple of classes that we created.
alpr_som.py: A neuron gas self-organizing map that was successful in implementation but we abandoned for a quicker method
recfinder.py: Our character segmentation algorithm. This depends on the image being preprocessed to black and white as mentioned in our conference paper.